Upgrade loader UI

+++ loader/Main/DataStore.cs
@@ -1,17 +1,35 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Windows;
 
 namespace PenguLoader.Main
 {
     static class DataStore
     {
-        public static void Dump()
+        public static bool IsDataStore(string path)
+        {
+            if (string.IsNullOrEmpty(path))
+                return false;
+
+            return Path.GetFileName(path).Equals("datastore", StringComparison.OrdinalIgnoreCase)
+                && File.Exists(path);
+        }
+
+        public static void DumpDataStore(string path)
         {
             try
             {
-                var output = Config.DataStorePath + ".d";
-                var bytes = File.ReadAllBytes(Config.DataStorePath);
+                var output = Path.GetTempFileName();
+                var bytes = File.ReadAllBytes(path);
+
+                if (bytes.Length == 0)
+                {
+                    MessageBox.Show("Your DataStore is empty!",
+                        Program.Name, MessageBoxButton.OK, MessageBoxImage.Information);
+                    return;
+                }
+
                 File.WriteAllBytes(output, Transform(bytes));
 
                 Process.Start(new ProcessStartInfo
@@ -23,6 +41,8 @@ namespace PenguLoader.Main
             }
             catch
             {
+                MessageBox.Show("Failed to dump your DataStore.\nPlease make sure the input is correct.",
+                    Program.Name, MessageBoxButton.OK, MessageBoxImage.Information);
             }
         }
